scripts/requirements.txt
nltk
scikit-learn
py-rouge
pandas
statsmodels
numpy
torch
transformers
spacy